+++ src/components/submit.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import _ from 'lodash';
 import classNames from 'classnames';
 
 export default React.createClass({
@@ -7,6 +8,7 @@ export default React.createClass({
     propTypes: {
         className: React.PropTypes.string,
         disableIfInvalid: React.PropTypes.bool,
+        onClick: React.PropTypes.func,
     },
 
     contextTypes: {
@@ -16,7 +18,18 @@ export default React.createClass({
     getDefaultProps() {
         return {
             disableIfInvalid: false,
+            onClick: _.identity,
         };
+    },
+
+    onClick(event) {
+        if (!this.context.form.isHtmlForm()) {
+            event.preventDefault();
+            event.stopPropagation();
+            this.context.form.submit();
+        }
+
+        this.props.onClick(event);
     },
 
     render() {
@@ -24,10 +37,12 @@ export default React.createClass({
 
         return (
             <input
+                {..._.omit(this.props, 'children')}
                 type="submit"
+                onClick={this.onClick}
                 className={classNames(this.props.className, { _disabled: disabled })}
                 disabled={this.props.disableIfInvalid && disabled}
-                value={this.props.children}/>
+                value={this.props.value || this.props.children} />
         );
     },
 });