+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.1.0
+* Form: Add nested forms support #16
+* Form: Add new prop `useHtmlForm` with default true value
+* Form: Add `submit`, `validate`, `isHtmlForm` methods into form child context
+* FormComponentMixin: Add processKeyPress helper for form components
+
 ## 2.0.5
 * Fixed imports